Point and Vector were moved from physics namespace to algebra namespace

+++ src/simulator/physics/types.h
@@ -6,7 +6,7 @@
 namespace types
 {
 typedef int obj_id;
-typedef float value_t;
+typedef double value_t;
 typedef long int duration_t;
 } // namespace types
